Switch to playwright nightly to fix FF on Debian 12
It appears when playwright officially added support for Debian 12
(https://github.com/microsoft/playwright/issues/23532) they also
started throwing an error when trying to run unsupported browsers
on it (which Firefox was in the latest release). This broke our CI
(even though Firefox ran just fine on Debian 12 for the past year).

Here we simply switch to playwright nightly, which we'll just stick
with for a while.