[C#] Move C# side to a TypeScript-based array/string conversion
Using the marshallers properly will be a ton of additional work,
and its not clear we can do it in the C -> C# direction anyway.
Instead, we can trivially just use the TypeScript conversion logic
which converts everything through pointers anyway.

We'd previously done this on the C side of things, but here we do
it on the C# side as well.