[C#] Use an instance of the callback delegate, not the fn itself
When passing the callback delegate to C from C# during static init,
it appears (but only on Windows) this actually creates a C# class
instance for the delegate and passes that. However, that class
instance is eventually GC'd and our callbacks will start failing.

Instead, we create a static instance of of the delegate and pass
that to C.