Opportunistically clone arguments when possible
Once we're generating code here we don't know whether an argument
passed by reference is actually something the callee holds on to
as a reference or if it just needs it temporarily. This means we
have to add a `refs_to` pointer to the passed object to the method
owner, potentially holding on to objects long after they're needed.

In one specific case, `ChannelManager::pay_for_bolt11_invoice`
holds on to the passed `Bolt11Invoice`, potentially leaking memory
on every paid invoice.

Sadly there's not a great fix for this in that we don't want to
take the less conservative stance and assume the passed object
isn't held on to (though it would very likely work in practice!).
Instead, here, we bite the bullet and just clone everything we can
as we pass it into Rust, avoiding the reference entirely.